requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.6.2.post1
attrs==23.1.0
azure-core==1.32.0
azure-identity==1.19.0
beautifulsoup4==4.12.3
blis==0.7.11
Brotli==1.1.0
bs4==0.0.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cryptography==43.0.3
cymem==2.0.8
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.9
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
filelock==3.13.1
frozenlist==1.5.0
fuzzywuzzy==0.18.0
gender-guesser==0.4.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.6
importlib_metadata==8.5.0
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
Levenshtein==0.25.1
MarkupSafe==2.1.3
microsoft-kiota-abstractions==1.6.0
microsoft-kiota-authentication-azure==1.6.0
microsoft-kiota-http==1.6.0
microsoft-kiota-serialization-form==1.6.0
microsoft-kiota-serialization-json==1.6.0
microsoft-kiota-serialization-multipart==1.6.0
microsoft-kiota-serialization-text==1.6.0
msal==1.31.0
msal-extensions==1.2.0
msgraph-core==1.1.6
msgraph-sdk==1.11.0
multidict==6.1.0
murmurhash==1.0.10
nlp==0.4.0
nltk==3.8.1
numpy==1.26.2
openpyxl==3.1.2
opentelemetry-api==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-semantic-conventions==0.49b0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
pendulum==3.0.0
pillow==10.4.0
pipenv==2024.4.0
platformdirs==4.3.6
portalocker==2.10.1
preshed==3.0.9
propcache==0.2.0
pyarrow==15.0.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.9.0
pypng==0.20220715.0
PySocks==1.7.1
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
pytz==2023.3.post1
pywin32==308
qrcode==7.4.2
rapidfuzz==3.9.1
regex==2023.12.25
reportlab==4.2.5
requests==2.31.0
selenium==4.26.1
setuptools==69.0.3
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
std-uritemplate==1.0.6
thinc==8.2.2
time-machine==2.16.0
toml==0.10.2
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
Unidecode==1.3.7
urllib3==2.2.3
virtualenv==20.28.0
wasabi==1.1.2
weasel==0.3.4
websocket-client==1.8.0
wrapt==1.16.0
wsproto==1.2.0
XlsxWriter==3.1.9
xxhash==3.4.1
yarl==1.17.1
zipp==3.20.2
